+++ database/migrations/2018_10_05_110109_add_href_test_to_services_table.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddHrefTestToServicesTable extends Migration
+{
+  /**
+   * Run the migrations.
+   *
+   * @return void
+   */
+  public function up()
+  {
+    Schema::table('services', function (Blueprint $table) {
+      $table->string('href_test')->nullable();
+    });
+  }
+
+  /**
+   * Reverse the migrations.
+   *
+   * @return void
+   */
+  public function down()
+  {
+    Schema::table('services', function (Blueprint $table) {
+      $table->dropColumn('href_test');
+    });
+  }
+}