+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ---
 
+## [1.1.0](https://github.com/MUTUAL-DE-SERVICIOS-AL-POLICIA/PVA-EV/tree/1.1.0) - 2018-10-05
+
+### Added
+
+* Link to go to test versión of each service
+
+### Changed
+
+* Spaces and margins to view more expanded services cards
+
 ## [1.0.0](https://github.com/MUTUAL-DE-SERVICIOS-AL-POLICIA/PVA-EV/tree/1.0.0) - 2018-10-04
 
 ### Added